Reorganized code for generating the lookup json files.
